+++ src/main/java/semillero/ecosistema/service/contracts/ProvinceService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface ProvinceService {
     List<ProvinceEntity> getAll();
     ProvinceEntity getByNombre(String nombre);
-    List<ProvinceEntity> getByPais(int pais_id);
+    List<ProvinceEntity> getByPais(Long pais_id);
     ProvinceEntity save(ProvinceEntity provinceEntity);
     ProvinceEntity update(String nameProvince, ProvinceEntity provinceEntity);
